Fix player dispose in undo reset when player rejoins the server

+++ AsyncWorldEdit-APIInner/src/main/java/org/primesoft/asyncworldedit/api/inner/ICron.java
@@ -59,7 +59,7 @@ public interface ICron {
      * @param entry
      * @param delayTime 
      */
-    void scheduleSessionForRemoval(IPlayerEntry entry, int delayTime);
-    
-    
+    void scheduleSessionForRemoval(IPlayerEntry entry, long delayTime);
+
+
 }
